Remove agrovoc linked data authority
The URL for the AgroVoc API has changed twice in the last year and a half.  It is undesirable to have to update QA for the changing API requirements.  The AgroVoc configuration will continue to be maintained at https://github.com/LD4P/linked_data_authorities/tree/master/qa_agrovoc.