Add Qa config for a default language for flitering and sorting
The default language can be set to nil to have all languages returned.

This modifies the filter algorithm to include literals not matching the preferred language in the case where there are no literals matching the preferred language for a predicate.  Better to return the non-preferred language value than no value at all.