add ability to encode the variable value
Encoding is primarily used to encode URIs that are part of the URL path instead of being a parameter.  But the generalized processing allows any variable to be marked as requiring URL encoding.

Also making all methods private unless they are needed in other classes.  In this refactor, that effect #default and #required?