Move search and find tests SearchQuery and FindTerm specs
These were tested in GenericAuthority because typically they are called as a delegate in this class.  But it is easier to read the tests when they are separated out to those for search and those for find.  So they were moved to the specs for classes that implement the search and find methods.