update tests to use real values found in the data
The particular stubbing that is happening when using graph filtering in QA does not seem to mind if bogus data is used.  But as we move to ldpath, it is looking for real data in the graph.

This is prep work for moving findterm to use ldpath.

Forces an update of CCI cache since fixtures change to update linked data authority configs.