Raise an exception if data can’t be normalized
Closes #247

In this case, the data can’t be normalized because the requested ID does not match the ID in the graph for the term.  This prevents the extraction of the URI from the graph based on the ID.

With this fix, an exception is raised when it happens so the error can be reported back to users.  This fails more gracefully than the previous failure which just returned empty data for uri, id, and label.