add a request id to the search and find request headers
When there is a heavy load of requests, logging can be interspersed for the various requests.  It is difficult to tell which log message applies to which request.  Adding a request id allows each log message to be prefixed with the id.

The request was also added to the request_header to allow downstream apps to have access to the request if they need additional information.