pin Faraday gem to < 2.0
Pinning Faraday to postpone upgrade to 2.0.  ActiveFedora currently pins to `’~> 0.12’` which is more strict, but compatible with the current pin in QA to `’< 2.0’`.

The 2.0 upgrade requires a configuration that is not supported in earlier versions.  See [upgrade doc](https://github.com/lostisland/faraday/blob/main/UPGRADING.md) for more information.