Fix test of raw serialized JSON
In some versions of ruby/dependencies, there is a literal `&` in raw serialized JSON source; in other versions, it's a `\u0026` (which should mean the same thing once parsed). We don't understand why this differs between different ruby/rails versions, or why it started differing at some point in past with no changes to local code.

But it seems safe to have the test not care whether the serialized JSON contains `&` or `\u0026` -- should be semantically equivalent once parsed. So change the test code to test against parsed value rather than raw value.