Move Qa::LinkedData::AuthorityService to lib/
This class is accessed in an *initializer* (see config/initializers/linked_data_authorities.rb:). In Rails 7, this is somehow a loading problem, it results in:

```
questioning_authority/config/initializers/linked_data_authorities.rb:1:in `<top (required)>': uninitialized constant Qa::LinkedData (NameError)
	from /Users/jrochkind/.gem/ruby/3.0.4/gems/railties-7.0.4/lib/rails/engine.rb:667:in `load'
```

Moving this class from app/ to lib/, and explicitly requiring it at boot in the engine.rb file, seems to resolve the problem and allow tests to pass on Rails 7. See also #351