Extracting `Qa.authority_for`
Prior to this commit, we used different initialization logic for linked
data authorities and non-linked data.  Further, the initialization logic
was locked away inside of a controller.

With this commit, we expose a method that provides consistent
initialization logic; albeit with different returned classes that have
slightly different implementation nuances.  The application can then
rely on that and we can share initialization logic across controllers
and also expose a method that allows non-controller initialization to
begin to follow the same logic pathway.

Where are the tests?  The initializations are covered by controller
tests, so I have chosen not to write new ones.

There is further work to do because LinkedData and non-LinkedData there
are different signatures for `find` and `search`.  Ideally, we would
normalize the public facing implementation logic.  However, this commit
preserves backwards compatibility, simply introducing a new feature.