Introducing `Qa::AuthorityWrapper`
Prior to this commit, the returned value from `Qa.authority_for` was
something that was inconsistent.  Depending on the controller (or authority), the `find`
and `search` would require different parameters.

With this commit, those nuances are encapsulated in their own methods.
This begins to work towards a more foundational promise of questioning
authority; namely that the end point that asked questions of QA didn't
need to know which type of authority it was.

That promise was partially delivered at the routes level but this commit
pushes that down to the internal API level.

As a bonus, and perhaps confusion, I've included the
`Qa::AuthorityRequestContext` as an interface to help convey how one
might provide a different (non-controller) context to instantiating and
authority.