Adding a `Qa::Authorities::Loc.type_for`
Prior to this commit, we always assumed that each "find" would request
from `https://id.loc.gov/authorities`.

With this commit, we use the "subauthority" to inform what the base URL
is for the `https://id.loc.gov/` host.  In some cases this is
`authorities` and in others it's as the code indicates.

Hopefully works to close the following:

- scientist-softserv/utk-hyku#267